+++ libs/tabularasa/src/thds/tabularasa/__main__.py
@@ -233,6 +233,8 @@ def print_data_diff_summaries(
     *,
     exit_code: bool = False,
     verbose: bool = False,
+    value_detail: bool = False,
+    value_detail_min_count: int = 0,
     heading_level: int = 0,
     tablefmt: str = diff_summary.DEFAULT_TABLEFMT,
     floatfmt: str = diff_summary.DEFAULT_FLOATFMT,
@@ -243,6 +245,10 @@ def print_data_diff_summaries(
     :param exit_code: if True, exit with code 1 if there is a positive diff (similar to `git diff --exit-code`)
     :param verbose: if True, show detailed row change status counts; otherwise show only single-column
       change counts
+    :param value_detail: if True, show detailed value change counts; otherwise show only statistics of the
+      types of changes as determined by the `verbose` flag
+    :param value_detail_min_count: minimum number of instances of a specific value update to show value-level
+      detail for. No effect when `value_detail` is False
     :param heading_level: increase this to render smaller headings on the markdown sections
     :param tablefmt: the table format to use for the markdown tables, as understood by `tabulate`
     :param floatfmt: the float format to use for the markdown tables, as understood by `tabulate`
@@ -253,6 +259,8 @@ def print_data_diff_summaries(
             d_diff,
             table_name,
             verbose,
+            value_detail=value_detail,
+            value_detail_min_count=value_detail_min_count,
             heading_level=heading_level,
             tablefmt=tablefmt,
             floatfmt=floatfmt,